services/sentiment-nlp/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.1
kafka-python==2.0.2
vaderSentiment==3.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.7.0